pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sas-graph-code-conversion"
version = "0.1.0"
description = "Convert SAS statistical pipelines to R/Python via graph-based intermediate representation."
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "networkx>=3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
sgcc = "sgcc.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
